IGNITE-17871 Maven build fix
