--- modules/raft/src/integrationTest/java/org/apache/ignite/raft/server/counter/TestMessageGroup.java
@@ -1,32 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License. You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.ignite.raft.server.counter;
-
-import org.apache.ignite.network.annotations.MessageGroup;
-
-/**
- * Message group for tests.
- */
-@MessageGroup(groupType = Short.MAX_VALUE - 100, groupName = "ItTestRaftMessages")
-public interface TestMessageGroup {
-    /** Message type for {@link GetValueCommand}. */
-    short GET_VALUE_COMMAND = 1000;
-
-    /** Message type for {@link IncrementAndGetCommand}. */
-    short INCREMENT_AND_GET_COMMAND = 1001;
-}
