+++ modules/raft/src/testFixtures/java/org/apache/ignite/raft/messages/TestMessageGroup.java
@@ -18,10 +18,17 @@
 package org.apache.ignite.raft.messages;
 
 import org.apache.ignite.network.annotations.MessageGroup;
+import org.apache.ignite.raft.server.counter.GetValueCommand;
+import org.apache.ignite.raft.server.counter.IncrementAndGetCommand;
 
 /**
  * Message group for tests.
  */
 @MessageGroup(groupType = 4, groupName = "TestRaftMessages")
 public interface TestMessageGroup {
+    /** Message type for {@link GetValueCommand}. */
+    short GET_VALUE_COMMAND = 1000;
+
+    /** Message type for {@link IncrementAndGetCommand}. */
+    short INCREMENT_AND_GET_COMMAND = 1001;
 }
