+++ modules/raft/src/testFixtures/java/org/apache/ignite/raft/server/counter/IncrementAndGetCommand.java
@@ -17,15 +17,23 @@
 
 package org.apache.ignite.raft.server.counter;
 
-import org.apache.ignite.internal.raft.ReadCommand;
+import org.apache.ignite.internal.raft.WriteCommand;
 import org.apache.ignite.network.annotations.Transferable;
+import org.apache.ignite.raft.messages.TestMessageGroup;
+import org.apache.ignite.raft.messages.TestRaftMessagesFactory;
 
 /**
- * Get a value command.
+ * Increment and get command.
  */
-@Transferable(TestMessageGroup.GET_VALUE_COMMAND)
-public interface GetValueCommand extends ReadCommand {
-    static GetValueCommand getValueCommand() {
-        return new ItTestRaftMessagesFactory().getValueCommand().build();
+//TODO Move to integration test directory when Maven build is not supported anymore.
+@Transferable(TestMessageGroup.INCREMENT_AND_GET_COMMAND)
+public interface IncrementAndGetCommand extends WriteCommand {
+    /**
+     * Returns the delta.
+     */
+    long delta();
+
+    static IncrementAndGetCommand incrementAndGetCommand(long delta) {
+        return new TestRaftMessagesFactory().incrementAndGetCommand().delta(delta).build();
     }
 }
